pyproject.toml
[project]
name = "function-api-calling"
version = "0.4.0"
description = "Тесты и функции для API"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.14.0",
    "pyyaml>=6.0.2",
    "rich>=14.2.0",
    "pytest>=7.0.0",
    "dotenv>=0.9.9",
    "requests>=2.32.5",
    "orjson>=3.11.7",
    "pytest-coverage>=0.0",
    "allure-pytest==2.15.2",
    "allure-python-commons==2.15.2",
]

[project.optional-dependencies]
dev = [
    "black>=25.11.0",
    "ruff>=0.14.8",
    "pytest-mock>=3.10.0",
    "mypy_extensions>=1.1.0",
    "typing_extensions>=4.15.0",
]


[tool.setuptools.packages.find]
where = ["src", "configurates"]

[tool.black]
line-length = 96
target-version = ['py312']
skip-string-normalization = false
include = '\.pyi?$'

[tool.ruff]
# Максимальная длина строки (совпадает с black)
line-length = 100
target-version = "py312"
fix = true

# Исключаем лишние папки
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "migrations",
    "__pycache__",
    "test_results",
    "snapshots"
]

# Включаем основные правила:
# E — ошибки стиля (PEP8)
# F — ошибки компиляции (flake8)
# I — импорты (isort)
# B — best practices (bugbear)
# UP — modern Python (pyupgrade)
preview = true

# Игнорируем правила, которые конфликтуют с black
[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "E203"]


